working on uniform cost search
need to change Path. It won’t be sufficient for how uniform cost search
needs to keep track of things.

added MapSet which should hopefully provide efficient support for doing
things that UCS requires.

on top of that we have Weights, which is the frontier type for a
Uniform Cost Search.